test-1/requirements.txt
pyTelegramBotAPI
scapy
websocket-client
telebot